+++ src/core/di/index.ts
@@ -1,4 +1,4 @@
 /**
  * DI 시스템의 진입점 (Barrel Export)
  */
-export * from './container';
+export * from './container.js';